+++ Core/DataStructures/mpi_buffers.jl
@@ -3,16 +3,16 @@
 mutable struct GhostBuffers
     lefts::CuVector{Particle}
     rights::CuVector{Particle}
-    counters::CuVector{Int32}
-    overflow_flag::CuVector{Int32}
-    buf_lengths::Int32
+    counters::CuVector{Int64}
+    overflow_flag::CuVector{Int64}
+    buf_lengths::Int64
 end #struct
-GhostBuffers(max_particles::Union{Int32,Int64}) = GhostBuffers(
+GhostBuffers(max_particles::Union{Int64,Int64}) = GhostBuffers(
     CuVector{Particle}(undef, max_particles),
     CuVector{Particle}(undef, max_particles),
-    CUDA.zeros(Int32, 2),
-    CUDA.zeros(Int32, 1),
-    Int32(max_particles)
+    CUDA.zeros(Int64, 2),
+    CUDA.zeros(Int64, 1),
+    Int64(max_particles)
 )
 
 
@@ -22,40 +22,40 @@ mutable struct MigrantBuffers
     stayers::CuVector{Particle}
     lefts::CuVector{Particle}
     rights::CuVector{Particle}
-    counters::CuVector{Int32}
-    overflow_flag::CuVector{Int32}
-    buf_lengths::Int32
-    stayer_overflow_flag::CuVector{Int32}
-    stayer_buf_length::Int32
+    counters::CuVector{Int64}
+    overflow_flag::CuVector{Int64}
+    buf_lengths::Int64
+    stayer_overflow_flag::CuVector{Int64}
+    stayer_buf_length::Int64
 end #struct
 MigrantBuffers(
-    max_particles_per_rank::Union{Int32,Int64},
-    max_sendrecv_particles::Union{Int32,Int64}
+    max_particles_per_rank::Union{Int64,Int64},
+    max_sendrecv_particles::Union{Int64,Int64}
 ) = MigrantBuffers(
     CuVector{Particle}(undef, max_particles_per_rank),
     CuVector{Particle}(undef, max_sendrecv_particles),
     CuVector{Particle}(undef, max_sendrecv_particles),
-    CUDA.zeros(Int32, 3),
-    CUDA.zeros(Int32, 1),
-    Int32(max_sendrecv_particles),
-    CUDA.zeros(Int32, 1),
-    Int32(max_particles_per_rank)
+    CUDA.zeros(Int64, 3),
+    CUDA.zeros(Int64, 1),
+    Int64(max_sendrecv_particles),
+    CUDA.zeros(Int64, 1),
+    Int64(max_particles_per_rank)
 )
 
 
 # --------- MPI Send/Recv Buffers --------- #
 
 mutable struct SendRecvBuffers
-    send_left::CuVector{UInt32}
-    send_right::CuVector{UInt32}
-    recv_left::CuVector{UInt32}
-    recv_right::CuVector{UInt32}
-    buf_lengths::Int32
+    send_left::CuVector{UInt64}
+    send_right::CuVector{UInt64}
+    recv_left::CuVector{UInt64}
+    recv_right::CuVector{UInt64}
+    buf_lengths::Int64
 end #struct
-SendRecvBuffers(max_particles::Union{Int32,Int64}) = SendRecvBuffers(
-    CuVector{UInt32}(undef, 4 * max_particles),
-    CuVector{UInt32}(undef, 4 * max_particles),
-    CuVector{UInt32}(undef, 4 * max_particles),
-    CuVector{UInt32}(undef, 4 * max_particles),
-    Int32(4 * max_particles)
+SendRecvBuffers(max_particles::Union{Int64,Int64}) = SendRecvBuffers(
+    CuVector{UInt64}(undef, 4 * max_particles),
+    CuVector{UInt64}(undef, 4 * max_particles),
+    CuVector{UInt64}(undef, 4 * max_particles),
+    CuVector{UInt64}(undef, 4 * max_particles),
+    Int64(4 * max_particles)
 )
